Cannonまで見ました

+++ ACG/ACG/ObjBullet.cpp
@@ -28,21 +28,21 @@ CObjBullet::CObjBullet(float x, float y)
 
 	//マウスの位置と主人公の位置からマウスの角度を求める
 	//マウスの位置情報取得
-	double mous_x = Input::GetPosX();
-	double mous_y = Input::GetPosY();
+	float mous_x = Input::GetPosX();
+	float mous_y = Input::GetPosY();
 
 	//主人公の位置からマウスの位置のベクトル情報取得
-	double vector_x = mous_x - (x - objmap->GetScrollX());
-	double vector_y = mous_y - (y - objmap->GetScrollY());
+	float vector_x = mous_x - (x - objmap->GetScrollX());
+	float vector_y = mous_y - (y - objmap->GetScrollY());
 
 	//斜辺取得
-	double hypotenuse = sqrt(vector_y * vector_y + vector_x * vector_x);
+	float hypotenuse = sqrt(vector_y * vector_y + vector_x * vector_x);
 
 	//角度を求める
 	m_r = acos(vector_x / hypotenuse);
 	//角度方向に移動
 	m_vx = cos(m_r) * m_speed;
-	m_r = m_r * 180.0 / 3.14;
+	m_r = m_r * 180.0f / 3.14f;
 
 	//マウスのY位置が主人公のY位置より下だったら
 	if (mous_y > y )
